Move global utility methods into their own file
